Add some uncomitted changes.
* Update changelog to include recent improvements.
* Update default messages to point to ghostpp.com.
* Add bot_mapgametype to default.cfg.
* Fix issue with EXEVersionHash on some platforms.

+++ ghost/bncsutilinterface.cpp
@@ -72,9 +72,9 @@ bool CBNCSUtilInterface :: HELP_SID_AUTH_CHECK( bool TFT, string war3Path, strin
 		getExeInfo( FileWar3EXE.c_str( ), (char *)&buf, 1024, (uint32_t *)&EXEVersion, BNCSUTIL_PLATFORM_X86 );
 		m_EXEInfo = buf;
 		m_EXEVersion = UTIL_CreateByteArray( EXEVersion, false );
-		uint32_t EXEVersionHash;
+		unsigned long EXEVersionHash;
 		checkRevisionFlat( valueStringFormula.c_str( ), FileWar3EXE.c_str( ), FileStormDLL.c_str( ), FileGameDLL.c_str( ), extractMPQNumber( mpqFileName.c_str( ) ), (unsigned long *)&EXEVersionHash );
-		m_EXEVersionHash = UTIL_CreateByteArray( EXEVersionHash, false );
+		m_EXEVersionHash = UTIL_CreateByteArray( (uint32_t) EXEVersionHash, false );
 		m_KeyInfoROC = CreateKeyInfo( keyROC, UTIL_ByteArrayToUInt32( clientToken, false ), UTIL_ByteArrayToUInt32( serverToken, false ) );
 
 		if( TFT )
